+++ src/components/AuthorLabel.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { getAuthorShortName } from '../util';
 
 export default function AuthorLabel({
