+++ src/components/CopyAuthorSecretButton.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { useCurrentAuthor } from '../hooks';
 import CopyButton from './_CopyButton';
