+++ src/components/EarthstarPeer.tsx
@@ -1,0 +1,68 @@
+import * as React from 'react';
+import { AuthorKeypair, IStorage } from 'earthstar';
+import LiveSyncer from './LiveSyncer';
+import {
+  CurrentAuthorContext,
+  CurrentWorkspaceContext,
+  IsLiveContext,
+  PubsContext,
+  StorageContext,
+} from '../contexts';
+
+export default function EarthstarPeer({
+  initWorkspaces = [],
+  initPubs = {},
+  initCurrentAuthor = null,
+  initCurrentWorkspace = null,
+  initIsLive = true,
+
+  children,
+}: {
+  initWorkspaces?: IStorage[];
+  initPubs?: Record<string, string[]>;
+  initCurrentAuthor?: AuthorKeypair | null;
+  initCurrentWorkspace?: string | null;
+  initIsLive?: boolean;
+  children: React.ReactNode;
+}) {
+  const [storages, setStorages] = React.useState(
+    initWorkspaces.reduce<Record<string, IStorage>>((acc, storage) => {
+      return { ...acc, [storage.workspace]: storage };
+    }, {})
+  );
+
+  const [pubs, setPubs] = React.useState(initPubs);
+
+  const [currentAuthor, setCurrentAuthor] = React.useState(initCurrentAuthor);
+
+  const [currentWorkspace, setCurrentWorkspace] = React.useState(
+    initCurrentWorkspace && storages[initCurrentWorkspace]
+      ? initCurrentWorkspace
+      : null
+  );
+  const [isLive, setIsLive] = React.useState(initIsLive);
+
+  return (
+    <StorageContext.Provider value={{ storages, setStorages }}>
+      <PubsContext.Provider value={{ pubs, setPubs }}>
+        <CurrentAuthorContext.Provider
+          value={{ currentAuthor, setCurrentAuthor }}
+        >
+          <CurrentWorkspaceContext.Provider
+            value={{ currentWorkspace, setCurrentWorkspace }}
+          >
+            <IsLiveContext.Provider value={{ isLive, setIsLive }}>
+              {children}
+              {Object.keys(storages).map(workspaceAddress => (
+                <LiveSyncer
+                  key={workspaceAddress}
+                  workspaceAddress={workspaceAddress}
+                />
+              ))}
+            </IsLiveContext.Provider>
+          </CurrentWorkspaceContext.Provider>
+        </CurrentAuthorContext.Provider>
+      </PubsContext.Provider>
+    </StorageContext.Provider>
+  );
+}
