+++ src/components/InvitationRedemptionForm.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { isErr } from 'earthstar';
 import Alert from '@reach/alert';
 import {
@@ -7,8 +7,7 @@ import {
   usePubs,
   useCurrentWorkspace,
 } from '../hooks';
-import { WorkspaceLabel } from '../components';
-import { EarthbarContext } from './earthbar';
+import WorkspaceLabel from '../components/WorkspaceLabel';
 
 export default function InvitationRedemptionForm({
   onRedeem,
@@ -20,7 +19,6 @@ export default function InvitationRedemptionForm({
   const [, setCurrentWorkspace] = useCurrentWorkspace();
   const [code, setCode] = React.useState('');
   const result = useInvitation(code);
-  const { setActiveIndex, setFocusedIndex } = React.useContext(EarthbarContext);
 
   const [uncheckedPubs, setUncheckedPubs] = React.useState<string[]>([]);
 
@@ -86,8 +84,6 @@ export default function InvitationRedemptionForm({
             setCode('');
             setUncheckedPubs([]);
             setCurrentWorkspace(workspace);
-            setActiveIndex(-1);
-            setFocusedIndex(-1);
           }
         }}
       >
