+++ src/components/SyncingCheckbox.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { useIsLive } from '../hooks';
 
 export default function SyncingCheckbox() {
