+++ src/components/WorkspaceLabel.tsx
@@ -9,7 +9,7 @@ Given a workspace address, returns an element which:
 
 */
 
-import React from 'react';
+import * as React from 'react';
 import { getWorkspaceName } from '../util';
 
 export default function WorkspaceLabel({
