+++ src/components/WorkspaceList.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 
 import { useWorkspaces } from '../hooks';
 import WorkspaceLabel from './WorkspaceLabel';
