+++ src/components/earthbar/index.ts
@@ -1,4 +1,7 @@
 export { default as AuthorTab } from './AuthorTab';
 export { default as WorkspaceTab } from './WorkspaceTab';
 export { default as MultiWorkspaceTab } from './MultiWorkspaceTab';
-export * from './Earthbar';
+export { default as Earthbar } from './Earthbar';
+export { default as EarthbarTab } from './EarthbarTab';
+export { default as EarthbarTabLabel } from './EarthbarTabLabel';
+export { default as EarthbarTabPanel } from './EarthbarTabPanel';
