+++ src/components/earthbar/types.ts
@@ -1,0 +1,7 @@
+import { Descendant } from '@reach/descendants';
+
+export type TabDescendant = Descendant<HTMLButtonElement> & {
+  disabled: boolean;
+};
+
+export type TabPanelDescendant = Descendant<HTMLDivElement>;
